pyproject.toml
[project]
name = "at-translate"
version = "0.1.0"
description = "An automatic translation tool based on OpenAI's GPT. Integrated with Weblate."
authors = [
    { name = "DragonChenTW", email = "teacher144123@gmail.com" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "openai>=1.30.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/at_translate"]
